connected_car_controller/requirements.txt
pyserial==3.0.1
pybluez==0.22
paver==1.3.4